Text2SQL/requirement.txt
chromadb==0.6.3
Flask==3.1.0
mysql_connector_repackaged==0.3.1
protobuf==5.29.3
pymongo==4.6.2